pyproject.toml
[project]
name = "telegram-attendance-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography==41.0.1",
    "openpyxl>=3.1.0",
    "pdfkit==1.0.0",
    "pillow>=12.1.0",
    "pytelegrambotapi==4.12.0",
    "python-dotenv==1.0.0",
    "python-telegram==0.18.0",
    "pytz==2023.3",
    "sqlalchemy==2.0.17",
]
